+++ src/OpenTelemetry.Exporter.InfluxDB/InfluxDBMetricsExporter.cs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using InfluxDB.Client;
-using InfluxDB.Client.Writes;
 using OpenTelemetry.Metrics;
 
 namespace OpenTelemetry.Exporter.InfluxDB;
@@ -11,34 +10,28 @@ internal sealed class InfluxDBMetricsExporter : BaseExporter<Metric>
 {
     private readonly IMetricsWriter writer;
     private readonly InfluxDBClient influxDbClient;
-    private readonly WriteApi writeApi;
+    private readonly WriteApiAsync writeApiAsync;
 
-    public InfluxDBMetricsExporter(IMetricsWriter writer, InfluxDBClient influxDbClient, WriteApi writeApi)
+    public InfluxDBMetricsExporter(IMetricsWriter writer, InfluxDBClient influxDbClient)
     {
         this.writer = writer;
         this.influxDbClient = influxDbClient;
-        this.writeApi = writeApi;
-
-        this.writeApi.EventHandler += (_, args) =>
-        {
-            switch (args)
-            {
-                case WriteErrorEvent writeErrorEvent:
-                    InfluxDBEventSource.Log.FailedToExport(writeErrorEvent.Exception.Message);
-                    break;
-                default:
-                    break;
-            }
-        };
+        this.writeApiAsync = influxDbClient.GetWriteApiAsync();
     }
 
     public override ExportResult Export(in Batch<Metric> batch)
     {
         try
         {
+            List<string> lineProtocol = [];
             foreach (var metric in batch)
             {
-                this.writer.Write(metric, this.ParentProvider?.GetResource(), this.writeApi);
+                this.writer.Write(metric, this.ParentProvider?.GetResource(), lineProtocol);
+            }
+
+            if (lineProtocol.Count > 0)
+            {
+                this.writeApiAsync.WriteRecordsAsync(lineProtocol).GetAwaiter().GetResult();
             }
 
             return ExportResult.Success;
@@ -54,7 +47,6 @@ internal sealed class InfluxDBMetricsExporter : BaseExporter<Metric>
     {
         if (disposing)
         {
-            this.writeApi.Dispose();
             this.influxDbClient.Dispose();
         }