+++ src/OpenTelemetry.Instrumentation.ConfluentKafka/InstrumentedProducer.cs
@@ -259,7 +259,11 @@ internal sealed class InstrumentedProducer<TKey, TValue> : IProducer<TKey, TValu
         produceException.Error.Code.ToString();
 
     private static string FormatArgumentException(ArgumentException argumentException) =>
+#if NET
         argumentException.GetType().FullName!;
+#else
+        argumentException.GetType().FullName;
+#endif
 
     private static void GetTags(string topic, out TagList tags, int? partition = null, string? errorType = null)
     {
@@ -322,12 +326,12 @@ internal sealed class InstrumentedProducer<TKey, TValue> : IProducer<TKey, TValu
         var spanName = string.Concat(ConfluentKafkaCommon.SendOperationName, " ", topic);
 
         // Provide the attributes that can influence sampling decisions at span creation time
-        var initialTags = new ActivityTagsCollection
+        var initialTags = new TagList
         {
-            [SemanticConventions.AttributeMessagingDestinationName] = topic,
-            [SemanticConventions.AttributeMessagingOperationName] = ConfluentKafkaCommon.SendOperationName,
-            [SemanticConventions.AttributeMessagingOperationType] = ConfluentKafkaCommon.SendOperationType,
-            [SemanticConventions.AttributeMessagingSystem] = ConfluentKafkaCommon.KafkaMessagingSystem,
+            { SemanticConventions.AttributeMessagingDestinationName, topic },
+            { SemanticConventions.AttributeMessagingOperationName, ConfluentKafkaCommon.SendOperationName },
+            { SemanticConventions.AttributeMessagingOperationType, ConfluentKafkaCommon.SendOperationType },
+            { SemanticConventions.AttributeMessagingSystem, ConfluentKafkaCommon.KafkaMessagingSystem },
         };
 
         if (partition is { } partitionValue)