+++ src/OpenTelemetry.Instrumentation.Http/Implementation/HttpTagHelper.cs
@@ -20,12 +20,46 @@ internal static class HttpTagHelper
     /// <returns>Span uri value.</returns>
     public static string GetUriTagValueFromRequestUri(Uri uri, bool disableQueryRedaction)
     {
-        if (string.IsNullOrEmpty(uri.UserInfo) && disableQueryRedaction)
+        string? query = null;
+
+        if (string.IsNullOrEmpty(uri.UserInfo))
         {
-            return uri.OriginalString;
+            if (disableQueryRedaction)
+            {
+                return uri.OriginalString;
+            }
+
+            // Redaction only rewrites the query when it contains a '=', so avoid
+            // recreating a string that is equivalent to the original string otherwise.
+            // Non HTTP(S) schemes with no authority behave slightly differently
+            // for AbsoluteUri so they ignore this optimization to ensure they
+            // return the right value to the caller.
+            var scheme = uri.Scheme;
+
+            if (scheme == Uri.UriSchemeHttps || scheme == Uri.UriSchemeHttp)
+            {
+                query = uri.Query;
+
+                var indexOfEquals =
+#if NET
+                    query.IndexOf('=', StringComparison.Ordinal);
+#else
+                    query.IndexOf('=');
+#endif
+
+                if (indexOfEquals < 0)
+                {
+                    return uri.AbsoluteUri;
+                }
+            }
         }
 
-        var query = disableQueryRedaction ? uri.Query : RedactionHelper.GetRedactedQueryString(uri.Query);
+        query ??= uri.Query;
+
+        if (!disableQueryRedaction)
+        {
+            query = RedactionHelper.GetRedactedQueryString(query);
+        }
 
         return string.Concat(uri.Scheme, Uri.SchemeDelimiter, uri.Authority, uri.AbsolutePath, query, uri.Fragment);
     }