+++ src/OpenTelemetry.Instrumentation.SqlClient/Implementation/SqlEventSourceListener.netfx.cs
@@ -240,14 +240,7 @@ internal sealed class SqlEventSourceListener : EventListener
 
         if (activity != null && activity.IsAllDataRequested)
         {
-            foreach (var name in SqlTelemetryHelper.SharedTagNames)
-            {
-                var value = activity.GetTagItem(name);
-                if (value != null)
-                {
-                    tags.Add(name, value);
-                }
-            }
+            SqlTelemetryHelper.AddSharedTags(activity, ref tags);
         }
         else
         {