+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -2,6 +2,27 @@
 
 ## Unreleased
 
+* Validate `FlushInterval` to prevent out-of-range values and
+  zero-interval spin loops.
+  ([#4860](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4860))
+
+* Fixed exceptions thrown from a user-supplied `Enrich`
+  callback preventing the `Activity` from being stopped.
+  ([#4900](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4900))
+
+* Fixed exceptions thrown by Redis profiler callbacks after
+  instrumentation is disposed.
+  ([#4905](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4905))
+
+* Fixed `Baggage.Current` being empty in samplers, processors and the `Filter`
+  and `Enrich` callbacks for Redis command activities.
+  ([#4927](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4927))
+
+* Fixed the Redis command drain thread inheriting the execution context of the
+  thread that created the instrumentation, which leaked that context's baggage
+  onto every Redis command activity.
+  ([#4927](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4927))
+
 ## 1.17.0-beta.1
 
 Released 2026-Jul-17